requirements.txt
slackclient
python-telegram-bot